Scraper/requirements.txt
selenium
beautifulsoup4
mariadb
python_dotenv